Code Description
This program reads and ini file for credentials and database etc. 
It will put out displays for what it reads. 
It also opens files and formats strings.
It querries detail using inline SQL
It writes and output file when complete 
and returns a 0 for no errors during processing. 
It demonstrates syntax, structure and coding standards that I used at the time.